Pre-Analysis Data Processing

getGenotypes() performs pre-analysis data processing of PLINK formatted unphased genotype data, including removal of SNPs and isolates with high proportions of missing data and SNPs with low minor allele frequencies. It also calculates SNP allele frequencies from either the input dataset or a specified reference dataset.

getGenotypes(ped.map, reference.ped.map = NULL, maf = 0.01,
  isolate.max.missing = 0.1, snp.max.missing = 0.1, chromosomes = NULL,
  input.map.distance = "cM", reference.map.distance = "cM")

Arguments

ped.map
A list with 2 objects:

A data frame which contains the PLINK PED information. The first six columns of this data frame are:

Family ID (type "character", "numeric" or "integer")

Isolate ID (type "character", "numeric" or "integer")

Paternal ID (type "character", "numeric" or "integer")

Maternal ID (type "character", "numeric" or "integer")

Multiplicity of infection (MOI) (1 = single infection or haploid, 2 = multiple infections or diploid)

Phenotype (type "character", "numeric" or "integer")

where each row describes a single isolate. The IDs are alphanumeric: the combination of family and isolate ID should uniquely identify a sample. The paternal, maternal and phenotype columns are not used by isoRelate, however they are required for completeness of a standard pedigree and are typically filled with the numeric value zero. Columns 7 onwards are the isolate genotypes where the A and B alleles are coded as 1 and 2 respectively and missing data is coded as 0. All SNPs must have two alleles specified and each allele should be in a separate column. For example, the alleles in columns 7 and 8 correspond to the unphased genotypes of SNP 1 in the map file. For single infections, genotypes should be specified as homozygous. Either both alleles should be missing (i.e. 0) or neither. Column names are not required.

A data frame which contains the PLINK MAP information. This data frame contains exactly four columns of information:

Chromosome (type "character", "numeric" or "integer")

SNP identifier (type "character")

Genetic map distance (centi morgans, cM, or morgans, M) (type "numeric")

Base-pair position (type "numeric" or "integer")

where each row describes a single SNP Genetic map distance and base-pair positions are expected to be positive values. The MAP file must be ordered by increasing genetic map distance. SNP identifiers can contain any characters expect spaces or tabs; also you should avoid * symbols in the names. The MAP file must contain as many markers as are in the PED file. Column names are not required.

reference.ped.map
An optional list containing reference data used to calculate SNP allele frequencies. The list has 2 objects in the same format as ped.map. The default value is reference.ped.map=NULL and isoRelate will calculate the SNP allele frequencies from the input data. This is not recommended for small datasets or datasets of mixed populations.

maf
A numeric value denoting the smallest minor allele frequency allowed in the analysis. The default value is 0.01.

isolate.max.missing
A numeric value denoting the maximum proportion of missing data allowed for each isolate. The default value is 0.1.

snp.max.missing
A numeric value denoting the maximum proportion of missing data allowed for each SNP. The default value is 0.1.

chromosomes
A vector containing a subset of chromosomes to perform formatting on. The default value is chromosomes=NULL which will reformat all genotypes for all chromosomes in the MAP data frame.

input.map.distance
A character string of either "M" or "cM" denoting whether the genetic map distances in the input MAP data frame are in Morgans (M) or centi-Morgans (cM). The default is cM.

reference.map.distance
A character string of either "M" or "cM" denoting whether the genetic map distances in the reference MAP data frame are in Morgans (M) or centi-Morgans (cM). The default is cM.

Value

A list of two objects named pedigree and genotypes:

A pedigree containing the isolates that remain after filtering. The pedigree is the first six columns of the PED file and these columns are headed fid, iid, pid, mid, moi and aff respectively.

A data frame with the first five columns:

Chromosome (type "character", "numeric" or "integer")

SNP identifiers (type "character")

Genetic map distance (Morgans, M) (type "numeric")

Base-pair position (type "integer")

Population allele frequency (type "integer")

where each row describes a single SNP. These columns are headed chr, snp_id, pos_M, pos_bp and freq respectively. Columns 6 onwards contain the genotype data for each isolate, where a single column corresponds to a single isolate. These columns are labeled with merged family IDs and isolate IDs separated by a slash symbol (/).

See also

getIBDparameters and getIBDsegments.

Examples

# take a look at the data
str(png_pedmap)

#> List of 2
#>  $ :'data.frame':	38 obs. of  59268 variables:
#>   ..$ fid  : chr [1:38] "PNG" "PNG" "PNG" "PNG" ...
#>   ..$ iid  : chr [1:38] "PN0003-C" "PN0004-C" "PN0005-C" "PN0006-C" ...
#>   ..$ pid  : num [1:38] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ mid  : num [1:38] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ moi  : num [1:38] 1 1 1 1 2 1 1 2 1 1 ...
#>   ..$ aff  : num [1:38] 2 2 2 2 2 2 2 2 2 2 ...
#>   ..$ 1    : num [1:38] 1 1 1 0 1 1 1 1 1 1 ...
#>   ..$ 2    : num [1:38] 1 1 1 0 1 1 1 1 1 1 ...
#>   ..$ 3    : num [1:38] 0 1 1 0 1 1 1 1 1 1 ...
#>   ..$ 4    : num [1:38] 0 1 1 0 1 1 1 1 1 1 ...
#>   ..$ 5    : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 6    : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 7    : num [1:38] 1 1 1 1 1 1 2 1 1 1 ...
#>   ..$ 8    : num [1:38] 1 1 1 1 2 1 2 2 1 1 ...
#>   ..$ 9    : num [1:38] 1 1 1 2 1 1 1 1 1 1 ...
#>   ..$ 10   : num [1:38] 1 1 1 2 1 1 1 1 1 1 ...
#>   ..$ 11   : num [1:38] 1 1 0 2 1 1 1 1 1 1 ...
#>   ..$ 12   : num [1:38] 1 1 0 2 1 1 1 1 1 1 ...
#>   ..$ 13   : num [1:38] 1 1 0 1 1 1 1 1 2 1 ...
#>   ..$ 14   : num [1:38] 1 1 0 1 1 1 1 1 2 1 ...
#>   ..$ 15   : num [1:38] 1 1 0 1 1 1 1 1 1 0 ...
#>   ..$ 16   : num [1:38] 1 1 0 1 1 1 1 1 1 0 ...
#>   ..$ 17   : num [1:38] 1 1 0 1 1 1 1 1 1 1 ...
#>   ..$ 18   : num [1:38] 1 1 0 1 1 1 1 1 1 1 ...
#>   ..$ 19   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 20   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 21   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 22   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 23   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 24   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 25   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 26   : num [1:38] 1 1 1 1 2 1 1 1 1 1 ...
#>   ..$ 27   : num [1:38] 2 2 2 1 1 2 1 1 1 1 ...
#>   ..$ 28   : num [1:38] 2 2 2 1 1 2 1 1 1 1 ...
#>   ..$ 29   : num [1:38] 2 2 2 1 1 2 1 1 1 1 ...
#>   ..$ 30   : num [1:38] 2 2 2 1 1 2 1 1 1 1 ...
#>   ..$ 31   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 32   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 33   : num [1:38] 1 1 1 1 1 1 1 1 2 1 ...
#>   ..$ 34   : num [1:38] 1 1 1 1 1 1 1 1 2 1 ...
#>   ..$ 35   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 36   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 37   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 38   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 39   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 40   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 41   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 42   : num [1:38] 1 1 1 1 2 1 1 1 1 1 ...
#>   ..$ 43   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 44   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 45   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 46   : num [1:38] 1 1 1 1 1 1 1 2 1 1 ...
#>   ..$ 47   : num [1:38] 1 1 1 1 1 1 0 1 1 1 ...
#>   ..$ 48   : num [1:38] 1 1 1 1 2 1 0 1 1 1 ...
#>   ..$ 49   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 50   : num [1:38] 1 1 1 1 2 1 1 1 1 1 ...
#>   ..$ 51   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 52   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 53   : num [1:38] 1 1 1 1 1 1 2 1 1 1 ...
#>   ..$ 54   : num [1:38] 1 1 1 1 2 1 2 2 1 1 ...
#>   ..$ 55   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 56   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 57   : num [1:38] 2 2 2 1 1 2 2 2 2 1 ...
#>   ..$ 58   : num [1:38] 2 2 2 1 2 2 2 2 2 1 ...
#>   ..$ 59   : num [1:38] 2 2 2 1 1 2 2 2 2 1 ...
#>   ..$ 60   : num [1:38] 2 2 2 1 2 2 2 2 2 1 ...
#>   ..$ 61   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 62   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 63   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 64   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 65   : num [1:38] 1 1 1 1 1 1 2 1 1 1 ...
#>   ..$ 66   : num [1:38] 1 1 1 1 2 1 2 1 1 1 ...
#>   ..$ 67   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 68   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 69   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 70   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 71   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 72   : num [1:38] 1 1 1 1 2 1 1 1 1 1 ...
#>   ..$ 73   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 74   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 75   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 76   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 77   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 78   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 79   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 80   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 81   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 82   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 83   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 84   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 85   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 86   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 87   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 88   : num [1:38] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ 89   : num [1:38] 1 2 1 1 1 1 1 1 1 1 ...
#>   ..$ 90   : num [1:38] 1 2 1 1 1 1 1 1 1 1 ...
#>   ..$ 91   : num [1:38] 1 2 1 1 1 1 1 1 1 1 ...
#>   ..$ 92   : num [1:38] 1 2 1 1 1 1 1 1 1 1 ...
#>   ..$ 93   : num [1:38] 0 1 1 1 2 1 0 1 2 2 ...
#>   .. [list output truncated]
#>  $ :'data.frame':	29631 obs. of  4 variables:
#>   ..$ chr   : chr [1:29631] "Pf3D7_01_v3" "Pf3D7_01_v3" "Pf3D7_01_v3" "Pf3D7_01_v3" ...
#>   ..$ snp_id: chr [1:29631] "Pf3D7_01_v3:93824" "Pf3D7_01_v3:94044" "Pf3D7_01_v3:94976" "Pf3D7_01_v3:95839" ...
#>   ..$ pos_cM: num [1:29631] 5.47 5.49 5.54 5.59 5.61 ...
#>   ..$ pos_bp: int [1:29631] 93824 94044 94976 95839 96128 96136 97100 97387 97718 98042 ...

# reformat and filter to call genotypes
my_genotypes <- getGenotypes(ped.map = png_pedmap,
                             reference.ped.map = NULL,
                             maf = 0.01,
                             isolate.max.missing = 0.1,
                             snp.max.missing = 0.1,
                             chromosomes = NULL,
                             input.map.distance = "cM",
                             reference.map.distance = "cM")

#> Begin filtering of 38 isolates and 29631 SNPs...
#> 19720 SNPs remain after MAF removal...
#> 18270 SNPs remain after missingness removal...

#> Warning: isolates removed due to genotype missingness: PNG/PN0041-C

#> 37 isolates remain after missingness removal...