Interval Cluster Networks

getIBDiclusters() produces a network of clusters of isolates that have been inferred IBD over a specified interval. Isolates that are not IBD over the interval are not included in the network or output. The networks are created using the R package igraph.

getIBDiclusters(ped.genotypes, ibd.segments, interval = NULL, prop = 0,
  hi.clust = FALSE)

Arguments

ped.genotypes
A list containing 2 objects. See the Value description in getGenotypes for more details on this input.

ibd.segments
A data frame containing the IBD segments detected by isoRelate. See the Value description in getIBDsegments for more details on this input.

interval
A vector of length 3 containing the region to identify clusters over. This vector should contain the chromosome ID, the start of the interval in base-pairs and the end of the interval in base-pairs; in this order respectively.

prop
Numeric value between 0 and 1 (inclusive). The minimum proportion of an interval (in base-pairs) shared IBD between a pair of isolates in order for the pair to be included in the network. For example, if prop=1 then two isolates will be included if they are IBD over the entire interval, whereas prop=0.5 will include isolates that are IBD over at least 50% of the interval. The default is prop=0, which includes isolates if they have an IBD segment that overlaps the interval in any way.

hi.clust
Logical. Whether to perform hierarchical clustering using the fastgreedy.community approach in the igraph package.

Value

A list of three objects named clusters, i.network and hi.clust:

A list where each object contains the names of isolates that form a disjoint cluster in the network. If hierarchical clustering has been performed then the clusters may not be disjoint.

An igraph network used in the construction of network plots. See http://igraph.org/r/ doe more details.

Logical. Whether or not hierarchical clustering has been performed.

See also

getGenotypes, getIBDsegments and getIBDpclusters.

Examples

# generate the isolates who are IBD over the Plasmodium falciparum CRT gene
my_i_clusters <- getIBDiclusters(ped.genotypes = png_genotypes,
                                 ibd.segments = png_ibd,
                                 interval = c("Pf3D7_07_v3", 403222, 406317),
                                 prop=0,
                                 hi.clust = FALSE)

#> Number of clusters = 5
#> Maximum cluster length = 9
#> Minimum cluster length = 2

str(my_i_clusters)

#> List of 3
#>  $ clusters :List of 5
#>   ..$ : chr [1:9] "PNG/PN0014-C" "PNG/PN0023-C" "PNG/PN0031-C" "PNG/PN0040-C" ...
#>   ..$ : chr [1:5] "PNG/PN0057-C" "PNG/PN0078-C" "PNG/PN0079-C" "PNG/PN0086-C" ...
#>   ..$ : chr [1:3] "PNG/PN0003-C" "PNG/PN0005-C" "PNG/PN0083-C"
#>   ..$ : chr [1:2] "PNG/PN0022-C" "PNG/PN0042-C"
#>   ..$ : chr [1:2] "PNG/PN0037-C" "PNG/PN0058-Cx"
#>  $ i.network:List of 10
#>   ..$ :List of 1
#>   .. ..$ PNG/PN0003-C:Class 'igraph.vs'  atomic [1:1] 2
#>   .. .. .. ..- attr(*, "env")=<weakref> 
#>   .. .. .. ..- attr(*, "graph")= chr "39a4976c-6a81-4e3a-86a3-5fef0eb92dbe"
#>   ..$ :List of 1
#>   .. ..$ PNG/PN0005-C:Class 'igraph.vs'  atomic [1:2] 1 17
#>   .. .. .. ..- attr(*, "env")=<weakref> 
#>   .. .. .. ..- attr(*, "graph")= chr "39a4976c-6a81-4e3a-86a3-5fef0eb92dbe"
#>   ..$ :List of 1
#>   .. ..$ PNG/PN0014-C:Class 'igraph.vs'  atomic [1:8] 5 6 8 9 10 11 13 18
#>   .. .. .. ..- attr(*, "env")=<weakref> 
#>   .. .. .. ..- attr(*, "graph")= chr "39a4976c-6a81-4e3a-86a3-5fef0eb92dbe"
#>   ..$ :List of 1
#>   .. ..$ PNG/PN0022-C:Class 'igraph.vs'  atomic [1:1] 19
#>   .. .. .. ..- attr(*, "env")=<weakref> 
#>   .. .. .. ..- attr(*, "graph")= chr "39a4976c-6a81-4e3a-86a3-5fef0eb92dbe"
#>   ..$ :List of 1
#>   .. ..$ PNG/PN0023-C:Class 'igraph.vs'  atomic [1:7] 3 6 8 9 10 11 13
#>   .. .. .. ..- attr(*, "env")=<weakref> 
#>   .. .. .. ..- attr(*, "graph")= chr "39a4976c-6a81-4e3a-86a3-5fef0eb92dbe"
#>   ..$ :List of 1
#>   .. ..$ PNG/PN0031-C:Class 'igraph.vs'  atomic [1:7] 3 5 8 9 10 11 13
#>   .. .. .. ..- attr(*, "env")=<weakref> 
#>   .. .. .. ..- attr(*, "graph")= chr "39a4976c-6a81-4e3a-86a3-5fef0eb92dbe"
#>   ..$ :List of 1
#>   .. ..$ PNG/PN0037-C:Class 'igraph.vs'  atomic [1:1] 20
#>   .. .. .. ..- attr(*, "env")=<weakref> 
#>   .. .. .. ..- attr(*, "graph")= chr "39a4976c-6a81-4e3a-86a3-5fef0eb92dbe"
#>   ..$ :List of 1
#>   .. ..$ PNG/PN0040-C:Class 'igraph.vs'  atomic [1:8] 3 5 6 9 10 11 13 18
#>   .. .. .. ..- attr(*, "env")=<weakref> 
#>   .. .. .. ..- attr(*, "graph")= chr "39a4976c-6a81-4e3a-86a3-5fef0eb92dbe"
#>   ..$ :List of 1
#>   .. ..$ PNG/PN0055-C:Class 'igraph.vs'  atomic [1:7] 3 5 6 8 10 11 13
#>   .. .. .. ..- attr(*, "env")=<weakref> 
#>   .. .. .. ..- attr(*, "graph")= chr "39a4976c-6a81-4e3a-86a3-5fef0eb92dbe"
#>   ..$ :List of 1
#>   .. ..$ PNG/PN0055-Cx:Class 'igraph.vs'  atomic [1:7] 3 5 6 8 9 11 13
#>   .. .. .. ..- attr(*, "env")=<weakref> 
#>   .. .. .. ..- attr(*, "graph")= chr "39a4976c-6a81-4e3a-86a3-5fef0eb92dbe"
#>   ..- attr(*, "class")= chr "igraph"
#>  $ hi.clust : logi FALSE