Genome Cluster Networks

getIBDpclusters() produces a network of clusters of isolates that share a minimum proportion of genome IBD. Isolates that do not share a minimum proportion IBD are not included in the network or output. The networks are created using R package igraph.

getIBDpclusters(ped.genotypes, ibd.segments, prop = 1, hi.clust = FALSE)

Arguments

ped.genotypes
A list containing 2 objects. See the Value description in getGenotypes for more details on this input.

ibd.segments
A data frame containing the IBD segments detected by isoRelate. See the Value description in getIBDsegments for more details on this input.

prop
Numeric value between (0,1]. The minimum proportion of genome shared IBD between a pair of isolates in order for the pair to be included in the network. For example, if prop=1 then two isolates will be included if they are IBD over the entire genome, i.e. identical isolates, whereas prop=0.5 will include isolates that share at least 50% of their genome IBD. The default is prop=1.

hi.clust
Logical. Whether to perform hierarchical clustering using the fastgreedy.community approach in the igraph package.

Value

A list of three objects named clusters, i.network and hi.clust:

A list where each object contains the names of isolates that form a disjoint cluster in the network. If hierarchical clustering has been performed then the clusters may not be disjoint.

An igraph network used in the construction of network plots. See http://igraph.org/r/ doe more details.

Logical. Whether or not hierarchical clustering has been performed.

See also

getGenotypes, getIBDsegments and getIBDiclusters.