Plot Cluster Networks

plotIBDclusters() Produces a figure of an isoRelate cluster network, where unique isolates are represented by vertices and a line is drawn between two vertices if the isolates have been inferred IBD via the criteria specified in either getIBDiclusters or getIBDpclusters.The networks are created using the R package igraph.

plotIBDclusters(ped.genotypes, clusters, groups = NULL, vertex.color = NULL,
  vertex.frame.color = "white", vertex.size = 4, vertex.name = FALSE,
  edge.color = "gray60", edge.width = 0.8, mark.border = "white",
  mark.col = "gray94", add.legend = TRUE, legend.x = -1.5,
  legend.y = -0.25, layout = NULL, return.layout = FALSE)

Arguments

ped.genotypes
A list containing 2 objects. See the Value description in getGenotypes for more details on this input.

clusters
A named list of three objects containing network information. See the Value description in either getIBDiclusters or getIBDpclusters for more details on this input.

groups
A data frame with 3 columns of information:

Family ID

Isolate ID

Group ID

Group ID, for example, can be the geographic regions where the isolates were collected. If groups is specified then each isolate in the pedigree must belong to a group. Vertices in the network will be colored according to group allocation. The default is groups=NULL and all vertices will have the same color.

vertex.color
A vector of characters or numeric values of the vertex colors in the network. If groups is specified then vertex.color should contain the same number of colors as unique groups.

vertex.frame.color
Character string or numeric value. A single color that will be used as the vertex border. Default is vertex.fame.color="white".

vertex.size
Numeric value indicating the size of the vertices in the network. Default is vertex.size=4.

vertex.name
Logical. Whether to add isolate names to the vertices. Default is vertex.name=FALSE.

edge.color
Character string or numeric value. A single color to be used for all edges. Default is edge.color="gray60".

edge.width
Numeric. A single value indicating the width of the edges. Default is edge.width=0.8.

mark.border
Character string or numeric value. A single color to be used for all borders in hierarchical clustering groups. Default is mark.border="white".

mark.col
Character string or numeric value. A single color to be used to fill hierarchical clustering groupings. Default is mark.col="gray94".

add.legend
Logical. Whether to include a legend in the plot. Default is add.legend=TRUE.

legend.x
Numerical. A single value indicating the x-coordinate of the legend, with default legend.x=-1.5.

legend.y
Numerical. A single value indicating the y-coordinate of the legend, with default legend.y=-0.25.

layout
A matrix containing the x and y coordinates of the vertices, generated using the Fruchterman-Reingold force-directed layout.

return.layout
Logical. Whether or not to return the layout matrix (vertex positions) in the network. This layout can be used as the input for the parameter layout to avoid different network configurations each time plotClusters() is run on the same network.

See also

getGenotypes, getIBDpclusters and getIBDiclusters.