Filtered Genotypes For The Papua New Guinea Dataset

Processed raw genotype data with the parameter settings as in the Vignette.

png_genotypes

Format

A list of two objects named pedigree and genotypes:

pedigree is a data frame with the following information:

Family ID

Isolate ID

Paternal ID. This is not used by isoRelate and is set to zero.

Maternal ID. This is not used by isoRelate and is set to zero.

Multiplicity of infection - 1 for single infection and 2 for multiple infection

Affection status of isolate. This is set to 2 and is ignored by isoRelate.

genotypes is a data frame with the first 5 columns:

Chromosome

SNP identifier

Genetic map distance

Base-pair position

Population allele frequency

where each row describes a single SNP. Columns 6 onwards contain the genotype data for each isolate, where a single column corresponds to a single isolate. These columns are labeled with merged family IDs and isolate IDs separated by a slash symbol (/).